Complete Change - New Foundation
After extensive testing of different systems, I think I finally have a
solid grasp on how to scrape the contents of a Wikipedia page.

There's other methods and libraries to do it but I did mine as bare
bones as possible.

In the future, I think I want to find a way to remove the footnotes so
the article doesn't have [1] [2] and so on next to different sentences.

This code is now the foundation going forward on this project. Will try
to commit often so I don't break.