Cleaned up input asks, can now word count based on input
The program will ask the user for a Wikipedia article + a word then dumb
out the entire article and the number of times their word appears.

Dumping out the article is probably unnecessary but this was just to
keep things clean

For future reference: The word count isn't TOTALLY true. It counts how
many times the word appears in the content of the articles. Any time the
word appears under a photo or in links in the footnotes or places
outside of the main article, it's not counted.

In other words, using Ctrl + F to count the number of times a word
appears in the article and comparing that to the output of my program
will likely almost never match.